% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoExpand.R
\name{autoExpand}
\alias{autoExpand}
\title{Data Frame Expansion Based on Unique Column Values}
\usage{
autoExpand(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
A wrapper function for \code{expand.grid()} that
automatically generates a combination of all
factors based upon the unique column entries in a
supplied data frame.
}
\details{
Test!
}
\examples{
data(HairEyeColor)
x <- as.data.frame(HairEyeColor)
autoExpand(x)
}

