% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct}
\alias{ct}
\title{Census Tract IDs and Area}
\format{A data frame with 1,450 rows and 2 columns:
\itemize{
   \item GEOID10: Census tract IDs from the US Census Bureau
   \item Area: Area of census tract in acres
   }}
\source{
TIGER/Line Shapefiles: \url{http://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
data(ct)
}
\description{
A simple table from the U.S. Census Bureau
that includes the ID for each census
tract in Washington State and the area (in acres)
of that census tract.
}
\details{
Note: Area was calculated using GIS software -
it is not in the same units as 'shape area' from the
US Census Bureau shapefiles.
}
\keyword{datasets}

