% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctpop}
\alias{ctpop}
\title{Census Tract Population}
\format{A data frame with 23,344 rows and 3 columns:
\itemize{
   \item GEOID10: Census tract IDs from the US Census Bureau
   \item Year: Year of estimate
   \item Population: Population estimate from the Washington
   State Office of Financial Management (OFM)
   }}
\source{
Office of Financial Management: \url{http://www.ofm.wa.gov/pop/smallarea/default.asp}
}
\usage{
data(ctpop)
}
\description{
This data set contains the estimated population of
Washington State census tracts from 2000 to 2015.
}
\details{
OFM uses a model of births, deaths, and migration to
forecast population size from the US Census Bureau data. These
forecasts are estimates, not actual counts of people. OFM
adjusts estimates, including previous years, in the fall of each
year.

Estimates in \code{ctpop} last updated: March 29, 2016
}
\keyword{datasets}

