% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{decode}
\alias{decode}
\title{Oracle-style Decode Function}
\usage{
decode(x, search, replace, default = NULL)
}
\arguments{
\item{x}{A vector}

\item{search}{Values to be replaced}

\item{replace}{Values to do the replacing}

\item{default}{Default value for entries unspecified in \code{search}}
}
\description{
Find and replace values by pairs. Includes an optional default 
for all 'left-over' values.
}
\details{
Credit for this function goes to flodel: 
\url{http://stackoverflow.com/questions/14105651/generate-column-values-with-multiple-conditions-in-r}
}

