% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negate Value Matching}
\usage{
... \%nin\% NA
}
\description{
\code{\%nin\%} is a binary operator which returns
a logical vector indicating when a match does not exist.
}
\examples{
x <- c(1, 3, 4)
y <- 6
y \%nin\% x
}

