% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepMax.R
\name{keepMax}
\alias{keepMax}
\title{De-duplicating while Retaining Max}
\usage{
keepMax(df, x, value, max = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{x}{A character vector of column names to remove on}

\item{value}{A character entry to sort on and keep the
largest (or smallest) value}

\item{max}{Optional parameter to keep max or min value. Defaults to
keeping max.}
}
\description{
This function returns unique entries from a data
frame with the highest value in a specified column of
interest.
}
\details{
This function is specifically for fulfilling common
data-call requests. An example is radon measurements
at residences each year. If a given residence
has more than one measurement in a given year, only the
most elevated test should be sent.
}
\examples{
data(iris)
x <- c("Petal.Length", "Petal.Width")
value <- "Sepal.Length"
keepMax(iris, x, value)
}

