% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{multMerge}
\alias{multMerge}
\title{Read and Merge Multiple Files}
\usage{
multMerge(mypath)
}
\arguments{
\item{mypath}{directory path}
}
\description{
Read in a directory path that contains multiple homogeneously 
formatted csv files to be merged into one data frame.
}
\details{
Credit for this function goes to Tony Cookson: 
\url{http://www.r-bloggers.com/merging-multiple-data-files-into-one-data-frame/}
}

