% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDupes.R
\name{rmDupes}
\alias{rmDupes}
\title{Removing Duplicate Rows}
\usage{
rmDupes(df, x)
}
\arguments{
\item{df}{A dataframe}

\item{x}{A character vector of column names to remove on}
}
\description{
A wrapper function for \code{duplicated()} that
finds and removes duplicate row entries from a data
frame.
}
\details{
This function will not match slight differences in names
and should not be used for addresses unless they have
already been corrected.
}
\examples{
data(iris)
x <- c("Petal.Length", "Petal.Width")
rmDupes(iris, x)
}

