% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDigit.R
\name{toDigit}
\alias{toDigit}
\title{Recode County Name to 1- and 2-digit codes}
\usage{
toDigit(x)
}
\arguments{
\item{x}{a character vector of WA state county names}
}
\description{
This function accepts county names in any case format and 
converts it to county code.
}
\examples{
x <- c('Adams', 'YAKIMA')
toDigit(x)
}

