% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toFIPS.R
\name{toFIPS}
\alias{toFIPS}
\title{Recode 1- and 2-digit County Code to FIPS}
\usage{
toFIPS(x, ct = TRUE)
}
\arguments{
\item{x}{a vector}

\item{ct}{option to convert to census tract or county. Defaults
to census tract}
}
\description{
This function accepts 1- and 2-digit county codes
and converts them either to county or census tract
FIPS.
}
\examples{
x <- c(3, 7, 63)
toFIPS(x, ct = FALSE)
}

