% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waName.R
\name{waName}
\alias{waName}
\title{Recode PWS ID}
\usage{
waName(x)
}
\arguments{
\item{x}{a character vector of PWS IDs}
}
\description{
This function accepts Public Water System (PWS) IDs and
outputs a modified form suitable for CDC data calls.
}
\examples{
x <- c("00029", "00045", "00050", "00119")
waName(x)
}

